requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
frozenlist==1.4.0
idna==3.4
joblib==1.3.1
multidict==6.0.4
mysql-connector-python==8.1.0
nltk==3.8.1
openai==0.27.8
protobuf==4.21.12
regex==2023.6.3
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
